refactor: move loggers into src folder #591
