+++ server/src/api/controllers/challenges.ts
@@ -4,7 +4,7 @@ import { ObjectId, Types } from "mongoose";
 
 import { IPaymentMethod, IUser, User } from "@models/User";
 
-import { logger } from "@logger";
+import { logger } from "src/loggers";
 import { FIELD_REQUIRED, NOT_FOUND, NOT_PERMITTED } from "../message";
 import { FIELD_INVALID } from "../message";
 import { Challenge, IChallenge } from "@models/Challenge";
