+++ server/src/loaders/index.ts
@@ -3,7 +3,7 @@ import passportLoader from "./passport";
 import expressLoader from "./express";
 import mongooseLoader from "./mongoose";
 import { loader as redisLoader } from "./redis";
-import { loader as loggerLoader } from "./logger";
+import loggerLoader from "./logger";
 import { configType } from "src/config/type";
 
 export default async (app: Express, config: configType) => {
