+++ server/src/loaders/logger.ts
@@ -1,16 +1,3 @@
-import { devLogger } from "log/devLogger";
-import { prodLogger } from "log/prodLogger";
+import { loader } from "src/loggers";
 
-let logger: any = null;
-
-const loader = (config: { NODE_ENV: string }) => {
-  if (config.NODE_ENV === "development") {
-    logger = devLogger;
-    console.log("✅ logger: devLogger");
-  } else if (config.NODE_ENV === "production") {
-    logger = prodLogger;
-    console.log("✅ logger: prodLogger");
-  }
-};
-
-export { loader, logger };
+export default loader;
