+++ server/tsconfig.json
@@ -13,8 +13,7 @@
     "baseUrl": ".",
     "paths": {
       "@models/*": ["models/*"],
-      "@controllers/*": ["src/api/controllers/*"],
-      "@logger": ["log/logger"]
+      "@controllers/*": ["src/api/controllers/*"]
     },
     "skipLibCheck": true
   },
